break up docker/kubernetes provider files into packages
